requirements.txt
Django>=5.2
djangorestframework>=3.16
django-allauth>=65.4
dj-rest-auth>=7.0
djangorestframework-simplejwt>=5.4
python-dotenv>=1.1
anthropic>=0.52
httpx>=0.28
stripe>=11.0
djangorestframework-camel-case>=1.4
requests>=2.31
